Delegate residual/ratio computation to the SegmentStarTable.
